Split Pedante evaluator contracts
Move Pedante response contracts and JSON parsing into focused package
modules while preserving the public `episodic.qa.pedante` API.

Keep evaluator orchestration in the package entry point and record the
tracking issue for the module decomposition. This removes the oversized
module and its broad `too-many-lines` suppression.
